+++ packages/column-views/src/carbon-isotopes/index.ts
@@ -1,3 +1,2 @@
 export * from "./isotopes-column";
 export * from "./data-provider";
-export * from "./core";