+++ packages/column-views/stories/sections.stories.ts
@@ -1,64 +1,15 @@
-import hyper from "@macrostrat/hyper";
-import styles from "./column.stories.module.sass";
 import { Meta, StoryObj } from "@storybook/react";
-import { useAPIResult } from "@macrostrat/ui-components";
 
-import {
-  Column,
-  LithologiesProvider,
-  ColoredUnitComponent,
-  preprocessSectionUnits,
-  preprocessUnits,
-} from "@macrostrat/column-views";
-import { Spinner } from "@blueprintjs/core";
-import { PatternProvider } from "@macrostrat/column-components/stories/base-section";
+import { ColoredUnitComponent } from "@macrostrat/column-views";
 import "@macrostrat/style-system";
-import { ColumnProps as BaseColumnProps } from "@macrostrat/column-views";
 import { ColumnAxisType } from "@macrostrat/column-components";
+import { StandaloneColumn, StandaloneColumnProps } from "./column-ui";
 
-const h = hyper.styled(styles);
+type Story = StoryObj<typeof StandaloneColumn>;
 
-interface ColumnProps extends BaseColumnProps {
-  id: number;
-}
-
-function useColumnUnits(col_id, inProcess) {
-  const status_code = inProcess ? "in process" : undefined;
-  return useAPIResult(
-    "https://macrostrat.org/api/v2/units",
-    { col_id, response: "long", status_code, show_position: true },
-    (res) => res.success.data
-  );
-}
-
-function useColumnBasicInfo(col_id, inProcess = false) {
-  const status_code = inProcess ? "in process" : undefined;
-  return useAPIResult(
-    "https://macrostrat.org/api/v2/columns",
-    { col_id, status_code },
-    (res) => {
-      return res.success.data[0];
-    }
-  );
-}
-
-function BasicSection(props: ColumnProps & { inProcess?: boolean }) {
-  const { id, inProcess, ...rest } = props;
-  const info = useColumnBasicInfo(id, inProcess);
-  const units = useColumnUnits(id, inProcess);
-
-  if (units == null || info == null) {
-    return h(Spinner);
-  }
-
-  return h("div", [h("h2", info.col_name), h(Column, { ...rest, units })]);
-}
-
-type Story = StoryObj<typeof BasicSection>;
-
-const meta: Meta<ColumnProps> = {
+const meta: Meta<StandaloneColumnProps> = {
   title: "Column views/Stratigraphic sections",
-  component: BasicSection,
+  component: StandaloneColumn,
   args: {
     id: 432,
     unconformityLabels: true,